Setup instructions: install Python 3.8
The Dropbox library is not compatible with Python 3.9.

Changing the Python version in the base conda environment is not ideal, so a conda environment just for running the project should be considered. However, I've had problems getting the `--use-conda` flag to work when starting Snakemake from a conda environment other than base.